TiVo Series 3 support:
Browsing more or less works, but transfers to a TiVo Series 3 (or TiVo HD)
caused the TiVo to crash. Examination of transfers of the same file using
tivohmo and pyTivo showed this difference:

Query (from TiVo):
    GET /TivoConnect?Command=QueryItem&Url=...

pyTivo:
    HTTP/1.1 404 Not Found

tivohmo:
    HTTP/1.1 200 OK
    [...headers omitted for brevity...]
    <?xml
      version="1.0"
      encoding="UTF-8"
      ?>
    <QueryItem/>

It is unclear to me if my TiVos hate all QueryItem responses, or just null
responses (like above). Since there is nothing being returned anyway, this
commit removes support for QueryItem.

The result is that tivohmo matches pyTivo behavior now and my transfers
are successful.